+++ include/roq/mbo_update.hpp
@@ -19,9 +19,9 @@ namespace roq {
 struct ROQ_PUBLIC MBOUpdate final {
   double price = NaN;              //!< Order price
   double remaining_quantity = {};  //!< Remaining order quantity
+  uint64_t priority = {};          //!< Queue priority
+  MBOOrderId order_id;             //!< Order identifier
   UpdateAction action = {};        //!< Order update action
-  uint32_t priority = {};          //!< Queue priority
-  ExternalTradeId order_id;        //!< Order identifier
 };
 
 }  // namespace roq
@@ -41,14 +41,14 @@ struct fmt::formatter<roq::MBOUpdate> {
         R"({{)"
         R"(price={}, )"
         R"(remaining_quantity={}, )"
-        R"(action={}, )"
         R"(priority={}, )"
-        R"(order_id="{}")"
+        R"(order_id="{}", )"
+        R"(action={})"
         R"(}})"_cf,
         value.price,
         value.remaining_quantity,
-        value.action,
         value.priority,
-        value.order_id);
+        value.order_id,
+        value.action);
   }
 };