+++ include/roq/string_types.hpp
@@ -15,6 +15,8 @@ static constexpr auto const MAX_LENGTH_EXCHANGE = size_t{32};
 static constexpr auto const MAX_LENGTH_SYMBOL = size_t{48};
 static constexpr auto const MAX_LENGTH_CURRENCY = size_t{32};
 
+static constexpr auto const MAX_LENGTH_MBO_ORDER_ID = size_t{36};  // note! UUID
+
 static constexpr auto const MAX_LENGTH_ORDER_TEMPLATE = size_t{16};
 static constexpr auto const MAX_LENGTH_EXTERNAL_ACCOUNT = size_t{64};
 static constexpr auto const MAX_LENGTH_EXTERNAL_ORDER_ID = size_t{64};
@@ -56,6 +58,10 @@ struct Symbol final : public String<detail::MAX_LENGTH_SYMBOL> {
 
 struct Currency final : public String<detail::MAX_LENGTH_CURRENCY> {
   using String<detail::MAX_LENGTH_CURRENCY>::String;
+};
+
+struct MBOOrderId final : public String<detail::MAX_LENGTH_MBO_ORDER_ID> {
+  using String<detail::MAX_LENGTH_MBO_ORDER_ID>::String;
 };
 
 struct OrderTemplate final : public String<detail::MAX_LENGTH_ORDER_TEMPLATE> {
@@ -118,6 +124,7 @@ static_assert(sizeof(Account) == detail::MAX_LENGTH_ACCOUNT);
 static_assert(sizeof(Exchange) == detail::MAX_LENGTH_EXCHANGE);
 static_assert(sizeof(Symbol) == detail::MAX_LENGTH_SYMBOL);
 static_assert(sizeof(Currency) == detail::MAX_LENGTH_CURRENCY);
+static_assert(sizeof(MBOOrderId) == detail::MAX_LENGTH_MBO_ORDER_ID);
 static_assert(sizeof(OrderTemplate) == detail::MAX_LENGTH_ORDER_TEMPLATE);
 static_assert(sizeof(ExternalAccount) == detail::MAX_LENGTH_EXTERNAL_ACCOUNT);
 static_assert(sizeof(ExternalOrderId) == detail::MAX_LENGTH_EXTERNAL_ORDER_ID);