+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* `MBOUpdate`: `priority` now `uint64_t`, `order_id` length 36 (UUID), re-ordering
+
 ## 0.9.1 &ndash; 2023-01-12
 
 ## 0.9.0 &ndash; 2022-12-22