+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes will be documented in this file.
 
 ## Head
 
+## 0.9.2 &ndash; 2023-02-22
+
 ### Added
 
 * `MBOUpdate.reason` + `UpdateReason`